IOS-2903 realy fixes: crash trying to create an account with port >= 32768 (only changed model before, not related code)

+++ SubProjects/MessageModel/MessageModel/MessageModel/Interface/VerifiableAccount/VerifiableAccount.swift
@@ -473,7 +473,7 @@ extension VerifiableAccount {
                         manuallyTrusted: Bool,
                         transport: ConnectionTransport) -> CdServer {
         server.address = address
-        server.port = Int16(port)
+        server.port = Int32(port)
         server.authMethod = authMethod?.rawValue
         server.serverType = serverType
         server.automaticallyTrusted = automaticallyTrusted
